requirements.txt
owlready2
rdflib
pyvis
python-docx
openai